refactor the geometric and ml files into separate files, change name of Layer to MultiImage
