+++ tests/test_multi_image.py
@@ -8,27 +8,29 @@ from jax import random, vmap
 TINY = 1.0e-5
 
 
-class TestLayer:
+class TestMultiImage:
 
     def testConstructor(self):
         key = random.PRNGKey(time.time_ns())
         D = 2
         N = 5
 
-        layer1 = geom.Layer({}, D, False)
-        assert layer1.D == D
-        assert layer1.is_torus == (False,) * D
-        for _, _ in layer1.items():
+        multi_image1 = geom.MultiImage({}, D, False)
+        assert multi_image1.D == D
+        assert multi_image1.is_torus == (False,) * D
+        for _, _ in multi_image1.items():
             assert False  # its empty, so this won't ever be called
 
         k = 0
-        layer2 = geom.Layer({k: random.normal(key, shape=((1,) + (N,) * D + (D,) * k))}, D, False)
-        assert layer2.D == D
-        assert layer2.is_torus == (False,) * D
-        assert layer2[0].shape == (1, N, N)
+        multi_image2 = geom.MultiImage(
+            {k: random.normal(key, shape=((1,) + (N,) * D + (D,) * k))}, D, False
+        )
+        assert multi_image2.D == D
+        assert multi_image2.is_torus == (False,) * D
+        assert multi_image2[0].shape == (1, N, N)
 
-        # layers can have multiple k values, and can have different size channels at each k
-        layer3 = geom.Layer(
+        # multi_images can have multiple k values, and can have different size channels at each k
+        multi_image3 = geom.MultiImage(
             {
                 0: random.normal(key, shape=((10,) + (N,) * D + (D,) * 0)),
                 1: random.normal(key, shape=((3,) + (N,) * D + (D,) * 1)),
@@ -36,17 +38,17 @@ class TestLayer:
             D,
             True,
         )
-        assert list(layer3.keys()) == [0, 1]
-        assert layer3[0].shape == (10, N, N)
-        assert layer3[1].shape == (3, N, N, D)
-        assert layer3.is_torus == (True,) * D
+        assert list(multi_image3.keys()) == [0, 1]
+        assert multi_image3[0].shape == (10, N, N)
+        assert multi_image3[1].shape == (3, N, N, D)
+        assert multi_image3.is_torus == (True,) * D
 
     def testCopy(self):
         key = random.PRNGKey(time.time_ns())
         D = 2
         N = 5
 
-        layer1 = geom.Layer(
+        multi_image1 = geom.MultiImage(
             {
                 0: random.normal(key, shape=((10,) + (N,) * D + (D,) * 0)),
                 1: random.normal(key, shape=((3,) + (N,) * D + (D,) * 1)),
@@ -55,17 +57,17 @@ class TestLayer:
             True,
         )
 
-        layer2 = layer1.copy()
-        assert layer1 is not layer2
+        multi_image2 = multi_image1.copy()
+        assert multi_image1 is not multi_image2
 
-        layer2[1] = jnp.arange(1 * (N**D) * D).reshape((1,) + (N,) * D + (D,) * 1)
-        assert layer2[1].shape == (1, N, N, D)
-        assert layer1[1].shape == (
+        multi_image2[1] = jnp.arange(1 * (N**D) * D).reshape((1,) + (N,) * D + (D,) * 1)
+        assert multi_image2[1].shape == (1, N, N, D)
+        assert multi_image1[1].shape == (
             3,
             N,
             N,
             D,
-        )  # original layer we copied from is unchanged
+        )  # original multi_image we copied from is unchanged
 
     def testFromImages(self):
         key = random.PRNGKey(time.time_ns())
@@ -74,19 +76,19 @@ class TestLayer:
 
         random_data = random.normal(key, shape=((10,) + (N,) * D + (D,) * 1))
         images = [geom.GeometricImage(data, 0, D) for data in random_data]
-        layer1 = geom.Layer.from_images(images)
-        assert layer1.D == D
-        assert layer1.is_torus == (True,) * D
-        assert list(layer1.keys()) == [(1, 0)]
-        assert layer1[(1, 0)].shape == (10, N, N, D)
+        multi_image1 = geom.MultiImage.from_images(images)
+        assert multi_image1.D == D
+        assert multi_image1.is_torus == (True,) * D
+        assert list(multi_image1.keys()) == [(1, 0)]
+        assert multi_image1[(1, 0)].shape == (10, N, N, D)
 
         # now images has multiple different values of k
         random_data2 = random.normal(key, shape=((33,) + (N,) * D + (D,) * 2))
         images.extend([geom.GeometricImage(data, 0, D) for data in random_data2])
-        layer2 = geom.Layer.from_images(images)
-        assert list(layer2.keys()) == [(1, 0), (2, 0)]
-        assert layer2[(1, 0)].shape == (10, N, N, D)
-        assert layer2[(2, 0)].shape == (33, N, N, D, D)
+        multi_image2 = geom.MultiImage.from_images(images)
+        assert list(multi_image2.keys()) == [(1, 0), (2, 0)]
+        assert multi_image2[(1, 0)].shape == (10, N, N, D)
+        assert multi_image2[(2, 0)].shape == (33, N, N, D, D)
 
     def testEq(self):
         key = random.PRNGKey(time.time_ns())
@@ -94,22 +96,22 @@ class TestLayer:
         N = 3
 
         key, subkey = random.split(key)
-        layer1 = geom.Layer(
+        multi_image1 = geom.MultiImage(
             {(0, 0): random.normal(subkey, shape=((10,) + (N,) * D + (D,) * 0))},
             D,
             True,
         )
-        layer1.append(1, 0, random.normal(subkey, shape=((10,) + (N,) * D + (D,) * 1)))
+        multi_image1.append(1, 0, random.normal(subkey, shape=((10,) + (N,) * D + (D,) * 1)))
 
-        layer2 = layer1.copy()
-        assert layer1 == layer2
+        multi_image2 = multi_image1.copy()
+        assert multi_image1 == multi_image2
 
         # keys do not match
-        layer3 = geom.Layer({(0, 0): jnp.ones((10,) + (N,) * D + (D,) * 0)}, D, True)
-        assert layer1 != layer3
+        multi_image3 = geom.MultiImage({(0, 0): jnp.ones((10,) + (N,) * D + (D,) * 0)}, D, True)
+        assert multi_image1 != multi_image3
 
         # values do not match
-        layer4 = geom.Layer(
+        multi_image4 = geom.MultiImage(
             {
                 (0,): jnp.ones((10,) + (N,) * D + (D,) * 0),
                 1: jnp.ones((10,) + (N,) * D + (D,) * 1),
@@ -117,18 +119,18 @@ class TestLayer:
             D,
             True,
         )
-        assert layer1 != layer4
+        assert multi_image1 != multi_image4
 
         # is_torus does not match
-        layer5 = geom.Layer(layer1.data, D, False)
-        assert layer1 != layer5
+        multi_image5 = geom.MultiImage(multi_image1.data, D, False)
+        assert multi_image1 != multi_image5
 
     def testAppend(self):
         key = random.PRNGKey(time.time_ns())
         D = 2
         N = 5
 
-        layer1 = geom.Layer(
+        multi_image1 = geom.MultiImage(
             {
                 (0, 0): random.normal(key, shape=((10,) + (N,) * D + (D,) * 0)),
                 (1, 0): random.normal(key, shape=((3,) + (N,) * D + (D,) * 1)),
@@ -138,33 +140,33 @@ class TestLayer:
         )
 
         image_block = random.normal(key, shape=((4,) + (N,) * D + (D,) * 1))
-        layer1.append(1, 0, image_block)
-        assert layer1[(0, 0)].shape == (10, N, N)  # unchanged
-        assert layer1[(1, 0)].shape == (7, N, N, D)  # updated 3+4=7
+        multi_image1.append(1, 0, image_block)
+        assert multi_image1[(0, 0)].shape == (10, N, N)  # unchanged
+        assert multi_image1[(1, 0)].shape == (7, N, N, D)  # updated 3+4=7
 
         image_block2 = random.normal(key, shape=((2,) + (N,) * D + (D,) * 2))
-        layer1.append(2, 0, image_block2)
-        assert layer1[(0, 0)].shape == (10, N, N)  # unchanged
-        assert layer1[(1, 0)].shape == (7, N, N, D)  # unchanged
-        assert layer1[(2, 0)].shape == (2, N, N, D, D)
+        multi_image1.append(2, 0, image_block2)
+        assert multi_image1[(0, 0)].shape == (10, N, N)  # unchanged
+        assert multi_image1[(1, 0)].shape == (7, N, N, D)  # unchanged
+        assert multi_image1[(2, 0)].shape == (2, N, N, D, D)
 
         # add an image block to the wrong k bucket
         with pytest.raises(AssertionError):
-            layer1.append(3, 0, image_block2)
+            multi_image1.append(3, 0, image_block2)
 
         # N is set by append if it is empty
-        layer2 = layer1.empty()
-        assert layer2.get_spatial_dims() is None
+        multi_image2 = multi_image1.empty()
+        assert multi_image2.get_spatial_dims() is None
 
-        layer2.append(0, 0, random.normal(key, shape=((10,) + (N,) * D + (D,) * 0)))
-        assert layer2.get_spatial_dims() == (N,) * D
+        multi_image2.append(0, 0, random.normal(key, shape=((10,) + (N,) * D + (D,) * 0)))
+        assert multi_image2.get_spatial_dims() == (N,) * D
 
     def testConcat(self):
         key = random.PRNGKey(time.time_ns())
         D = 2
         N = 5
 
-        layer1 = geom.Layer(
+        multi_image1 = geom.MultiImage(
             {
                 (0, 0): random.normal(key, shape=((10,) + (N,) * D + (D,) * 0)),
                 (1, 0): random.normal(key, shape=((3,) + (N,) * D + (D,) * 1)),
@@ -172,7 +174,7 @@ class TestLayer:
             D,
             True,
         )
-        layer2 = geom.Layer(
+        multi_image2 = geom.MultiImage(
             {
                 (1, 0): random.normal(key, shape=((4,) + (N,) * D + (D,) * 1)),
                 (2, 0): random.normal(key, shape=((5,) + (N,) * D + (D,) * 2)),
@@ -181,31 +183,31 @@ class TestLayer:
             True,
         )
 
-        layer3 = layer1.concat(layer2)
-        assert list(layer3.keys()) == [(0, 0), (1, 0), (2, 0)]
-        assert layer3[(0, 0)].shape == (10, N, N)
-        assert layer3[(1, 0)].shape == (7, N, N, D)
-        assert layer3[(2, 0)].shape == (5, N, N, D, D)
+        multi_image3 = multi_image1.concat(multi_image2)
+        assert list(multi_image3.keys()) == [(0, 0), (1, 0), (2, 0)]
+        assert multi_image3[(0, 0)].shape == (10, N, N)
+        assert multi_image3[(1, 0)].shape == (7, N, N, D)
+        assert multi_image3[(2, 0)].shape == (5, N, N, D, D)
 
         # mismatched D
-        layer4 = geom.Layer(
+        multi_image4 = geom.MultiImage(
             {(0, 0): random.normal(key, shape=((10,) + (N,) * D + (D,) * 0))}, D, True
         )
-        layer5 = geom.Layer(
+        multi_image5 = geom.MultiImage(
             {(0, 0): random.normal(key, shape=((10,) + (N,) * 3 + (D,) * 0))}, 3, True
         )
         with pytest.raises(AssertionError):
-            layer4.concat(layer5)
+            multi_image4.concat(multi_image5)
 
         # mismatched is_torus
-        layer6 = geom.Layer(
+        multi_image6 = geom.MultiImage(
             {(0, 0): random.normal(key, shape=((10,) + (N,) * D + (D,) * 0))}, D, True
         )
-        layer7 = geom.Layer(
+        multi_image7 = geom.MultiImage(
             {(0, 0): random.normal(key, shape=((10,) + (N,) * D + (D,) * 0))}, D, False
         )
         with pytest.raises(AssertionError):
-            layer6.concat(layer7)
+            multi_image6.concat(multi_image7)
 
     def testAdd(self):
         key = random.PRNGKey(time.time_ns())
@@ -217,7 +219,7 @@ class TestLayer:
             key, 9
         )
 
-        layer1 = geom.Layer(
+        multi_image1 = geom.MultiImage(
             {
                 (0, 0): random.normal(subkey1, shape=((channels,) + (N,) * D + (D,) * 0)),
                 (1, 0): random.normal(subkey2, shape=((channels,) + (N,) * D + (D,) * 1)),
@@ -226,7 +228,7 @@ class TestLayer:
             True,
         )
 
-        layer2 = geom.Layer(
+        multi_image2 = geom.MultiImage(
             {
                 (0, 0): random.normal(subkey3, shape=((channels,) + (N,) * D + (D,) * 0)),
                 (1, 0): random.normal(subkey4, shape=((channels,) + (N,) * D + (D,) * 1)),
@@ -235,45 +237,45 @@ class TestLayer:
             True,
         )
 
-        layer3 = layer1 + layer2
-        layer4 = geom.Layer(
+        multi_image3 = multi_image1 + multi_image2
+        multi_image4 = geom.MultiImage(
             {
-                (0, 0): layer1[(0, 0)] + layer2[(0, 0)],
-                (1, 0): layer1[(1, 0)] + layer2[(1, 0)],
+                (0, 0): multi_image1[(0, 0)] + multi_image2[(0, 0)],
+                (1, 0): multi_image1[(1, 0)] + multi_image2[(1, 0)],
             },
             D,
             True,
         )
 
-        assert layer3 == layer4
+        assert multi_image3 == multi_image4
 
-        # mismatched layer types
-        layer5 = geom.Layer(
+        # mismatched multi_image types
+        multi_image5 = geom.MultiImage(
             {(0, 0): random.normal(subkey5, shape=((channels,) + (N,) * D + (D,) * 0))},
             D,
             True,
         )
-        layer6 = geom.Layer(
+        multi_image6 = geom.MultiImage(
             {(1, 0): random.normal(subkey6, shape=((channels,) + (N,) * D + (D,) * 1))},
             D,
             True,
         )
         with pytest.raises(AssertionError):
-            layer5 + layer6
+            multi_image5 + multi_image6
 
         # mismatched number of channels
-        layer7 = geom.Layer(
+        multi_image7 = geom.MultiImage(
             {(0, 0): random.normal(subkey7, shape=((channels + 1,) + (N,) * D + (D,) * 0))},
             D,
             True,
         )
-        layer8 = geom.Layer(
+        multi_image8 = geom.MultiImage(
             {(0, 0): random.normal(subkey8, shape=((channels,) + (N,) * D + (D,) * 0))},
             D,
             True,
         )
         with pytest.raises(TypeError):
-            layer7 + layer8
+            multi_image7 + multi_image8
 
     def testMul(self):
         key = random.PRNGKey(0)
@@ -283,7 +285,7 @@ class TestLayer:
 
         key, subkey1, subkey2 = random.split(key, 3)
 
-        layer1 = geom.Layer(
+        multi_image1 = geom.MultiImage(
             {
                 (0, 0): random.normal(subkey1, shape=(channels,) + (N,) * D),
                 (1, 0): random.normal(subkey2, shape=(channels,) + (N,) * D + (D,)),
@@ -292,21 +294,21 @@ class TestLayer:
             True,
         )
 
-        layer2 = layer1 * 3
-        assert jnp.allclose(layer2[(0, 0)], layer1[(0, 0)] * 3)
-        assert jnp.allclose(layer2[(1, 0)], layer1[(1, 0)] * 3)
-        assert layer2.D == D
-        assert layer2.is_torus == (True,) * D
+        multi_image2 = multi_image1 * 3
+        assert jnp.allclose(multi_image2[(0, 0)], multi_image1[(0, 0)] * 3)
+        assert jnp.allclose(multi_image2[(1, 0)], multi_image1[(1, 0)] * 3)
+        assert multi_image2.D == D
+        assert multi_image2.is_torus == (True,) * D
 
-        layer3 = layer1 * -1
-        assert jnp.allclose(layer3[(0, 0)], layer1[(0, 0)] * -1)
-        assert jnp.allclose(layer3[(1, 0)], layer1[(1, 0)] * -1)
-        assert layer2.D == D
-        assert layer2.is_torus == (True,) * D
+        multi_image3 = multi_image1 * -1
+        assert jnp.allclose(multi_image3[(0, 0)], multi_image1[(0, 0)] * -1)
+        assert jnp.allclose(multi_image3[(1, 0)], multi_image1[(1, 0)] * -1)
+        assert multi_image2.D == D
+        assert multi_image2.is_torus == (True,) * D
 
-        # try to multiply two layers together
+        # try to multiply two multi_images together
         with pytest.raises(AssertionError):
-            layer1 * layer1
+            multi_image1 * multi_image1
 
     def testDiv(self):
         key = random.PRNGKey(0)
@@ -316,7 +318,7 @@ class TestLayer:
 
         key, subkey1, subkey2 = random.split(key, 3)
 
-        layer1 = geom.Layer(
+        multi_image1 = geom.MultiImage(
             {
                 (0, 0): random.normal(subkey1, shape=(channels,) + (N,) * D),
                 (1, 0): random.normal(subkey2, shape=(channels,) + (N,) * D + (D,)),
@@ -325,40 +327,40 @@ class TestLayer:
             True,
         )
 
-        layer2 = layer1 / 3
-        assert jnp.allclose(layer2[(0, 0)], layer1[(0, 0)] / 3)
-        assert jnp.allclose(layer2[(1, 0)], layer1[(1, 0)] / 3)
-        assert layer2.D == D
-        assert layer2.is_torus == (True,) * D
+        multi_image2 = multi_image1 / 3
+        assert jnp.allclose(multi_image2[(0, 0)], multi_image1[(0, 0)] / 3)
+        assert jnp.allclose(multi_image2[(1, 0)], multi_image1[(1, 0)] / 3)
+        assert multi_image2.D == D
+        assert multi_image2.is_torus == (True,) * D
 
-        layer3 = layer1 / -1
-        assert jnp.allclose(layer3[(0, 0)], layer1[(0, 0)] / -1)
-        assert jnp.allclose(layer3[(1, 0)], layer1[(1, 0)] / -1)
-        assert layer2.D == D
-        assert layer2.is_torus == (True,) * D
+        multi_image3 = multi_image1 / -1
+        assert jnp.allclose(multi_image3[(0, 0)], multi_image1[(0, 0)] / -1)
+        assert jnp.allclose(multi_image3[(1, 0)], multi_image1[(1, 0)] / -1)
+        assert multi_image2.D == D
+        assert multi_image2.is_torus == (True,) * D
 
-        # try to multiply two layers together
+        # try to multiply two multi_images together
         with pytest.raises(AssertionError):
-            layer1 * layer1
+            multi_image1 * multi_image1
 
     def testSize(self):
         D = 2
         N = 5
 
-        # empty layer
-        layer1 = geom.Layer({}, D)
-        assert layer1.size() == 0
+        # empty multi_image
+        multi_image1 = geom.MultiImage({}, D)
+        assert multi_image1.size() == 0
 
-        # basic scalar layer
-        layer2 = geom.Layer({(0, 0): jnp.ones((1,) + (N,) * D)}, D)
-        assert layer2.size() == N**D
+        # basic scalar multi_image
+        multi_image2 = geom.MultiImage({(0, 0): jnp.ones((1,) + (N,) * D)}, D)
+        assert multi_image2.size() == N**D
 
-        # layer with channels
-        layer3 = geom.Layer({(0, 0): jnp.ones((4,) + (N,) * D)}, D)
-        assert layer3.size() == (4 * N**D)
+        # multi_image with channels
+        multi_image3 = geom.MultiImage({(0, 0): jnp.ones((4,) + (N,) * D)}, D)
+        assert multi_image3.size() == (4 * N**D)
 
-        # more complex layer
-        layer4 = geom.Layer(
+        # more complex multi_image
+        multi_image4 = geom.MultiImage(
             {
                 (0, 0): jnp.ones((1,) + (N,) * D),
                 (1, 0): jnp.ones((4,) + (N,) * D + (D,)),
@@ -367,7 +369,7 @@ class TestLayer:
             },
             D,
         )
-        assert layer4.size() == (N**D + 4 * N**D * D + 2 * N**D * D + 3 * N**D * D * D)
+        assert multi_image4.size() == (N**D + 4 * N**D * D + 2 * N**D * D + 3 * N**D * D * D)
 
     def testVector(self):
         # Test the from_vector and to_vector functions
@@ -375,7 +377,7 @@ class TestLayer:
         D = 2
         N = 5
 
-        layer_example = geom.Layer(
+        multi_image_example = geom.MultiImage(
             {
                 (0, 0): jnp.ones((1,) + (N,) * D),
                 (1, 0): jnp.ones((1,) + (N,) * D + (D,)),
@@ -385,18 +387,18 @@ class TestLayer:
         )
 
         key, subkey = random.split(key)
-        rand_data = random.normal(subkey, shape=(layer_example.size(),))
+        rand_data = random.normal(subkey, shape=(multi_image_example.size(),))
 
-        rand_layer = geom.Layer.from_vector(rand_data, layer_example)
+        rand_multi_image = geom.MultiImage.from_vector(rand_data, multi_image_example)
 
-        assert rand_layer.size() == layer_example.size()
-        assert jnp.allclose(rand_layer.to_vector(), rand_data)
+        assert rand_multi_image.size() == multi_image_example.size()
+        assert jnp.allclose(rand_multi_image.to_vector(), rand_data)
 
-    def testToFromScalarLayer(self):
+    def testToFromScalarMultiImage(self):
         D = 2
         N = 5
 
-        layer_example = geom.Layer(
+        multi_image_example = geom.MultiImage(
             {
                 (0, 0): jnp.ones((1,) + (N,) * D),
                 (1, 0): jnp.ones((1,) + (N,) * D + (D,)),
@@ -405,15 +407,15 @@ class TestLayer:
             D,
         )
 
-        scalar_layer = layer_example.to_scalar_layer()
+        scalar_multi_image = multi_image_example.to_scalar_multi_image()
 
-        assert len(scalar_layer.keys()) == 1
-        assert next(iter(scalar_layer.keys())) == (0, 0)
-        assert jnp.allclose(scalar_layer[(0, 0)], jnp.ones((1 + D + D * D,) + (N,) * D))
+        assert len(scalar_multi_image.keys()) == 1
+        assert next(iter(scalar_multi_image.keys())) == (0, 0)
+        assert jnp.allclose(scalar_multi_image[(0, 0)], jnp.ones((1 + D + D * D,) + (N,) * D))
 
         key = random.PRNGKey(0)
         key, subkey1, subkey2, subkey3, subkey4 = random.split(key, 5)
-        rand_layer = geom.Layer(
+        rand_multi_image = geom.MultiImage(
             {
                 (0, 0): random.normal(subkey1, shape=((3,) + (N,) * D)),
                 (1, 0): random.normal(subkey2, shape=((1,) + (N,) * D + (D,))),
@@ -425,9 +427,11 @@ class TestLayer:
 
         layout = {(0, 0): 3, (1, 0): 1, (1, 1): 2, (2, 0): 1}
 
-        scalar_layer2 = rand_layer.to_scalar_layer()
-        assert list(scalar_layer2.keys()) == [(0, 0)]
-        assert rand_layer == rand_layer.to_scalar_layer().from_scalar_layer(layout)
+        scalar_multi_image2 = rand_multi_image.to_scalar_multi_image()
+        assert list(scalar_multi_image2.keys()) == [(0, 0)]
+        assert rand_multi_image == rand_multi_image.to_scalar_multi_image().from_scalar_multi_image(
+            layout
+        )
 
     def testTimesGroupElement(self):
         N = 5
@@ -436,12 +440,12 @@ class TestLayer:
         vmap_times_gg = vmap(geom.times_group_element, in_axes=(None, 0, None, None))
         key = random.PRNGKey(0)
         for D in [2, 3]:
-            layer = geom.Layer({}, D)
+            multi_image = geom.MultiImage({}, D)
 
             for parity in [0, 1]:
                 for k in [0, 1, 2, 3]:
                     key, subkey = random.split(key)
-                    layer.append(
+                    multi_image.append(
                         k,
                         parity,
                         random.normal(subkey, shape=((channels,) + (N,) * D + (D,) * k)),
@@ -450,11 +454,11 @@ class TestLayer:
             operators = geom.make_all_operators(D)
 
             for gg in operators:
-                rotated_layer = layer.times_group_element(gg)
+                rotated_multi_image = multi_image.times_group_element(gg)
 
-                for (k, parity), img_block in layer.items():
+                for (k, parity), img_block in multi_image.items():
                     rotated_block = vmap_times_gg(D, img_block, parity, gg)
-                    assert jnp.allclose(rotated_layer[(k, parity)], rotated_block)
+                    assert jnp.allclose(rotated_multi_image[(k, parity)], rotated_block)
 
     def testNorm(self):
         N = 5
@@ -465,7 +469,7 @@ class TestLayer:
 
         # norm of scalars, pseudo scalars, and vectors
         key, subkey1, subkey2, subkey3 = random.split(key, 4)
-        layer = geom.Layer(
+        multi_image = geom.MultiImage(
             {
                 (0, 0): random.normal(subkey1, shape=(channels,) + (N,) * D),
                 (0, 1): random.normal(subkey2, shape=(channels,) + (N,) * D),
@@ -474,15 +478,17 @@ class TestLayer:
             D,
         )
 
-        normed_layer = layer.norm()
-        assert list(normed_layer.keys()) == [(0, 0)]  # odd parity is converted to even parity
-        assert normed_layer[(0, 0)].shape == ((3 * channels,) + (N,) * D)
-        assert jnp.allclose(normed_layer[(0, 0)][:channels], jnp.abs(layer[(0, 0)]))
-        assert jnp.allclose(normed_layer[(0, 0)][channels : 2 * channels], jnp.abs(layer[(0, 1)]))
-        vector_norm = jnp.linalg.norm(
-            layer[(1, 0)].reshape(layer[(1, 0)].shape[: 1 + D] + (-1,)), axis=1 + D
+        normed_multi_image = multi_image.norm()
+        assert list(normed_multi_image.keys()) == [(0, 0)]  # odd parity is converted to even parity
+        assert normed_multi_image[(0, 0)].shape == ((3 * channels,) + (N,) * D)
+        assert jnp.allclose(normed_multi_image[(0, 0)][:channels], jnp.abs(multi_image[(0, 0)]))
+        assert jnp.allclose(
+            normed_multi_image[(0, 0)][channels : 2 * channels], jnp.abs(multi_image[(0, 1)])
         )
-        assert jnp.allclose(normed_layer[(0, 0)][2 * channels :], vector_norm)
+        vector_norm = jnp.linalg.norm(
+            multi_image[(1, 0)].reshape(multi_image[(1, 0)].shape[: 1 + D] + (-1,)), axis=1 + D
+        )
+        assert jnp.allclose(normed_multi_image[(0, 0)][2 * channels :], vector_norm)
 
     def testGetComponent(self):
         N = 5
@@ -491,62 +497,62 @@ class TestLayer:
         timesteps = 4
         key = random.PRNGKey(0)
         key, subkey1 = random.split(key, 2)
-        layer = geom.Layer(
+        multi_image = geom.MultiImage(
             {
                 (0, 0): random.normal(subkey1, shape=(channels * timesteps,) + (N,) * D),
             },
             D,
         )
-        assert isinstance(layer.get_component(0, future_steps=timesteps), geom.Layer)
+        assert isinstance(multi_image.get_component(0, future_steps=timesteps), geom.MultiImage)
         assert jnp.allclose(
-            layer.get_component(0, future_steps=timesteps)[(0, 0)],
-            layer[(0, 0)].reshape((-1, timesteps) + (N,) * D)[0],
+            multi_image.get_component(0, future_steps=timesteps)[(0, 0)],
+            multi_image[(0, 0)].reshape((-1, timesteps) + (N,) * D)[0],
         )
         assert jnp.allclose(
-            layer.get_component(1, future_steps=timesteps)[(0, 0)],
-            layer[(0, 0)].reshape((-1, timesteps) + (N,) * D)[1],
+            multi_image.get_component(1, future_steps=timesteps)[(0, 0)],
+            multi_image[(0, 0)].reshape((-1, timesteps) + (N,) * D)[1],
         )
 
         # slices work as well
         assert jnp.allclose(
-            layer.get_component(slice(0, 2), future_steps=timesteps)[(0, 0)],
-            layer[(0, 0)]
+            multi_image.get_component(slice(0, 2), future_steps=timesteps)[(0, 0)],
+            multi_image[(0, 0)]
             .reshape((-1, timesteps) + (N,) * D)[:2]
             .reshape((2 * timesteps,) + (N,) * D),
         )
 
 
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-# Test BatchLayer
+# Test Batchmulti_image
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 
-class TestBatchLayer:
+class TestBatchmulti_image:
 
     def testConstructor(self):
         key = random.PRNGKey(time.time_ns())
         D = 2
         N = 5
 
-        layer1 = geom.BatchLayer({}, D, False)
-        assert layer1.D == D
-        assert layer1.is_torus == (False,) * D
-        for _, _ in layer1.items():
+        multi_image1 = geom.BatchMultiImage({}, D, False)
+        assert multi_image1.D == D
+        assert multi_image1.is_torus == (False,) * D
+        for _, _ in multi_image1.items():
             assert False  # its empty, so this won't ever be called
 
         k = 0
-        layer2 = geom.BatchLayer(
+        multi_image2 = geom.BatchMultiImage(
             {(k, 0): random.normal(key, shape=((10, 1) + (N,) * D + (D,) * k))},
             D,
             False,
         )
-        assert layer2.D == D
-        assert layer2.is_torus == (False,) * D
-        assert layer2[(0, 0)].shape == (10, 1, N, N)
+        assert multi_image2.D == D
+        assert multi_image2.is_torus == (False,) * D
+        assert multi_image2[(0, 0)].shape == (10, 1, N, N)
 
-        # layers can have multiple k values with different channels,
+        # multi_images can have multiple k values with different channels,
         # but they should have same batch size, although this is currently unenforced
-        layer3 = geom.Layer(
+        multi_image3 = geom.MultiImage(
             {
                 (0, 0): random.normal(key, shape=((5, 10) + (N,) * D + (D,) * 0)),
                 (1, 0): random.normal(key, shape=((5, 3) + (N,) * D + (D,) * 1)),
@@ -554,10 +560,10 @@ class TestBatchLayer:
             D,
             True,
         )
-        assert list(layer3.keys()) == [(0, 0), (1, 0)]
-        assert layer3[(0, 0)].shape == (5, 10, N, N)
-        assert layer3[(1, 0)].shape == (5, 3, N, N, D)
-        assert layer3.is_torus == (True,) * D
+        assert list(multi_image3.keys()) == [(0, 0), (1, 0)]
+        assert multi_image3[(0, 0)].shape == (5, 10, N, N)
+        assert multi_image3[(1, 0)].shape == (5, 3, N, N, D)
+        assert multi_image3.is_torus == (True,) * D
 
     def testGetSubset(self):
         key = random.PRNGKey(time.time_ns())
@@ -565,64 +571,64 @@ class TestBatchLayer:
         N = 5
         k = 1
 
-        layer1 = geom.BatchLayer(
+        multi_image1 = geom.BatchMultiImage(
             {(k, 0): random.normal(key, shape=((100, 1) + (N,) * D + (D,) * k))},
             D,
             False,
         )
 
-        layer2 = layer1.get_subset(jnp.array([3]))
-        assert layer2.D == layer1.D
-        assert layer2.is_torus == layer1.is_torus
-        assert layer2.L == 1
-        assert layer2[(k, 0)].shape == (1, 1, N, N, D)
-        assert jnp.allclose(layer2[(k, 0)][0], layer1[(k, 0)][3])
+        multi_image2 = multi_image1.get_subset(jnp.array([3]))
+        assert multi_image2.D == multi_image1.D
+        assert multi_image2.is_torus == multi_image1.is_torus
+        assert multi_image2.L == 1
+        assert multi_image2[(k, 0)].shape == (1, 1, N, N, D)
+        assert jnp.allclose(multi_image2[(k, 0)][0], multi_image1[(k, 0)][3])
 
-        layer3 = layer1.get_subset(jnp.array([3, 23, 4, 17]))
-        assert layer3.L == 4
-        assert layer3[(k, 0)].shape == (4, 1, N, N, D)
-        assert jnp.allclose(layer3[(k, 0)], layer1[(k, 0)][jnp.array([3, 23, 4, 17])])
+        multi_image3 = multi_image1.get_subset(jnp.array([3, 23, 4, 17]))
+        assert multi_image3.L == 4
+        assert multi_image3[(k, 0)].shape == (4, 1, N, N, D)
+        assert jnp.allclose(multi_image3[(k, 0)], multi_image1[(k, 0)][jnp.array([3, 23, 4, 17])])
 
         # Indices must be a jax array
         with pytest.raises(AssertionError):
-            layer1.get_subset([3])
+            multi_image1.get_subset([3])
 
         with pytest.raises(AssertionError):
-            layer1.get_subset((0, 2, 3))
+            multi_image1.get_subset((0, 2, 3))
 
         with pytest.raises(AssertionError):
-            layer1.get_subset(jnp.array(0))
+            multi_image1.get_subset(jnp.array(0))
 
-    def testGetOneLayer(self):
+    def testGetOneBatchMultiImage(self):
         key = random.PRNGKey(time.time_ns())
         D = 2
         N = 5
         k = 1
 
-        layer1 = geom.BatchLayer(
+        multi_image1 = geom.BatchMultiImage(
             {(k, 0): random.normal(key, shape=((100, 1) + (N,) * D + (D,) * k))},
             D,
             False,
         )
 
-        layer2 = layer1.get_one_layer()
-        assert isinstance(layer2, geom.Layer)
-        assert layer2[(1, 0)].shape == (1, N, N, D)
-        assert jnp.allclose(layer1[(1, 0)][0], layer2[(1, 0)])
+        multi_image2 = multi_image1.get_one_batch_multi_image()
+        assert isinstance(multi_image2, geom.MultiImage)
+        assert multi_image2[(1, 0)].shape == (1, N, N, D)
+        assert jnp.allclose(multi_image1[(1, 0)][0], multi_image2[(1, 0)])
 
         idx = 12
-        layer3 = layer1.get_one_layer(idx)
-        assert isinstance(layer2, geom.Layer)
-        assert layer3[(1, 0)].shape == (1, N, N, D)
-        assert jnp.allclose(layer1[(1, 0)][idx], layer3[(1, 0)])
+        multi_image3 = multi_image1.get_one_batch_multi_image(idx)
+        assert isinstance(multi_image2, geom.MultiImage)
+        assert multi_image3[(1, 0)].shape == (1, N, N, D)
+        assert jnp.allclose(multi_image1[(1, 0)][idx], multi_image3[(1, 0)])
 
     def testAppend(self):
-        # For BatchLayer, append should probably only be used while it is vmapped to a Layer
+        # For Batchmulti_image, append should probably only be used while it is vmapped to a multi_image
         key = random.PRNGKey(time.time_ns())
         D = 2
         N = 5
 
-        layer1 = geom.Layer(
+        multi_image1 = geom.MultiImage(
             {
                 (0, 0): random.normal(key, shape=((5, 10) + (N,) * D + (D,) * 0)),
                 (1, 0): random.normal(key, shape=((5, 3) + (N,) * D + (D,) * 1)),
@@ -631,22 +637,26 @@ class TestBatchLayer:
             True,
         )
 
-        def mult(layer, param):
-            out_layer = layer.empty()
-            for (k, parity), image_block in layer.items():
-                out_layer.append(k, parity, param * jnp.ones(image_block.shape))
+        def mult(multi_image, param):
+            out_multi_image = multi_image.empty()
+            for (k, parity), image_block in multi_image.items():
+                out_multi_image.append(k, parity, param * jnp.ones(image_block.shape))
 
-            return out_layer
+            return out_multi_image
 
-        layer2 = vmap(mult)(layer1, jnp.arange(5))
-        assert layer2.D == layer1.D
-        assert layer2.is_torus == layer1.is_torus
-        assert layer2.keys() == layer1.keys()
-        for layer2_image, layer1_image, num in zip(layer2[(0, 0)], layer1[(0, 0)], jnp.arange(5)):
-            assert jnp.allclose(layer2_image, num * jnp.ones(layer1_image.shape))
+        multi_image2 = vmap(mult)(multi_image1, jnp.arange(5))
+        assert multi_image2.D == multi_image1.D
+        assert multi_image2.is_torus == multi_image1.is_torus
+        assert multi_image2.keys() == multi_image1.keys()
+        for multi_image2_image, multi_image1_image, num in zip(
+            multi_image2[(0, 0)], multi_image1[(0, 0)], jnp.arange(5)
+        ):
+            assert jnp.allclose(multi_image2_image, num * jnp.ones(multi_image1_image.shape))
 
-        for layer2_image, layer1_image, num in zip(layer2[(1, 0)], layer1[(1, 0)], jnp.arange(5)):
-            assert jnp.allclose(layer2_image, num * jnp.ones(layer1_image.shape))
+        for multi_image2_image, multi_image1_image, num in zip(
+            multi_image2[(1, 0)], multi_image1[(1, 0)], jnp.arange(5)
+        ):
+            assert jnp.allclose(multi_image2_image, num * jnp.ones(multi_image1_image.shape))
 
     def testConcat(self):
         key = random.PRNGKey(time.time_ns())
@@ -655,7 +665,7 @@ class TestBatchLayer:
 
         key, subkey1, subkey2, subkey3, subkey4 = random.split(key, 5)
 
-        layer1 = geom.BatchLayer(
+        multi_image1 = geom.BatchMultiImage(
             {
                 (1, 0): random.normal(subkey1, shape=((5, 10) + (N,) * D + (D,) * 1)),
                 (2, 0): random.normal(subkey2, shape=((5, 3) + (N,) * D + (D,) * 2)),
@@ -663,7 +673,7 @@ class TestBatchLayer:
             D,
             True,
         )
-        layer2 = geom.BatchLayer(
+        multi_image2 = geom.BatchMultiImage(
             {
                 (1, 0): random.normal(subkey3, shape=((7, 10) + (N,) * D + (D,) * 1)),
                 (2, 0): random.normal(subkey4, shape=((7, 3) + (N,) * D + (D,) * 2)),
@@ -672,51 +682,54 @@ class TestBatchLayer:
             True,
         )
 
-        layer3 = layer1.concat(layer2)
-        assert layer3.D == D
-        assert layer3.is_torus == (True,) * D
-        assert layer3[(1, 0)].shape == (12, 10, N, N, D)
-        assert layer3[(2, 0)].shape == (12, 3, N, N, D, D)
-        assert jnp.allclose(layer3[(1, 0)], jnp.concatenate([layer1[(1, 0)], layer2[(1, 0)]]))
+        multi_image3 = multi_image1.concat(multi_image2)
+        assert multi_image3.D == D
+        assert multi_image3.is_torus == (True,) * D
+        assert multi_image3[(1, 0)].shape == (12, 10, N, N, D)
+        assert multi_image3[(2, 0)].shape == (12, 3, N, N, D, D)
+        assert jnp.allclose(
+            multi_image3[(1, 0)], jnp.concatenate([multi_image1[(1, 0)], multi_image2[(1, 0)]])
+        )
 
         key, subkey5, subkey6 = random.split(key, 3)
-        layer4 = geom.BatchLayer(
+        multi_image4 = geom.BatchMultiImage(
             {(1, 0): random.normal(subkey5, shape=((5, 10) + (N,) * D + (D,) * 1))},
             D,
             True,
         )
-        layer5 = geom.BatchLayer(
+        multi_image5 = geom.BatchMultiImage(
             {(1, 0): random.normal(subkey6, shape=((5, 2) + (N,) * D + (D,) * 1))},
             D,
             True,
         )
 
-        layer6 = layer4.concat(layer5, axis=1)
-        assert layer6.D == D
-        assert list(layer6.keys()) == [(1, 0)]
-        assert layer6[(1, 0)].shape == (5, 12, N, N, D)
+        multi_image6 = multi_image4.concat(multi_image5, axis=1)
+        assert multi_image6.D == D
+        assert list(multi_image6.keys()) == [(1, 0)]
+        assert multi_image6[(1, 0)].shape == (5, 12, N, N, D)
         assert jnp.allclose(
-            layer6[(1, 0)], jnp.concatenate([layer4[(1, 0)], layer5[(1, 0)]], axis=1)
+            multi_image6[(1, 0)],
+            jnp.concatenate([multi_image4[(1, 0)], multi_image5[(1, 0)]], axis=1),
         )
 
     def testSize(self):
         D = 2
         N = 5
 
-        # empty layer
-        layer1 = geom.BatchLayer({}, D)
-        assert layer1.size() == 0
+        # empty multi_image
+        multi_image1 = geom.BatchMultiImage({}, D)
+        assert multi_image1.size() == 0
 
-        # basic scalar layer
-        layer2 = geom.BatchLayer({(0, 0): jnp.ones((1, 1) + (N,) * D)}, D)
-        assert layer2.size() == N**D
+        # basic scalar multi_image
+        multi_image2 = geom.BatchMultiImage({(0, 0): jnp.ones((1, 1) + (N,) * D)}, D)
+        assert multi_image2.size() == N**D
 
-        # layer with channels
-        layer3 = geom.BatchLayer({(0, 0): jnp.ones((2, 4) + (N,) * D)}, D)
-        assert layer3.size() == (2 * 4 * N**D)
+        # multi_image with channels
+        multi_image3 = geom.BatchMultiImage({(0, 0): jnp.ones((2, 4) + (N,) * D)}, D)
+        assert multi_image3.size() == (2 * 4 * N**D)
 
-        # more complex layer
-        layer4 = geom.BatchLayer(
+        # more complex multi_image
+        multi_image4 = geom.BatchMultiImage(
             {
                 (0, 0): jnp.ones((3, 1) + (N,) * D),
                 (1, 0): jnp.ones((3, 4) + (N,) * D + (D,)),
@@ -725,14 +738,14 @@ class TestBatchLayer:
             },
             D,
         )
-        assert layer4.size() == (3 * (N**D + 4 * N**D * D + 2 * N**D * D + 3 * N**D * D * D))
+        assert multi_image4.size() == (3 * (N**D + 4 * N**D * D + 2 * N**D * D + 3 * N**D * D * D))
 
-    def testToFromScalarLayer(self):
+    def testToFromScalarMultiImage(self):
         D = 2
         N = 5
 
         L = 1
-        layer_example = geom.BatchLayer(
+        multi_image_example = geom.BatchMultiImage(
             {
                 (0, 0): jnp.ones((L, 1) + (N,) * D),
                 (1, 0): jnp.ones((L, 1) + (N,) * D + (D,)),
@@ -741,16 +754,16 @@ class TestBatchLayer:
             D,
         )
 
-        scalar_layer = layer_example.to_scalar_layer()
+        scalar_multi_image = multi_image_example.to_scalar_multi_image()
 
-        assert len(scalar_layer.keys()) == 1
-        assert next(iter(scalar_layer.keys())) == (0, 0)
-        assert jnp.allclose(scalar_layer[(0, 0)], jnp.ones((1, 1 + D + D * D) + (N,) * D))
+        assert len(scalar_multi_image.keys()) == 1
+        assert next(iter(scalar_multi_image.keys())) == (0, 0)
+        assert jnp.allclose(scalar_multi_image[(0, 0)], jnp.ones((1, 1 + D + D * D) + (N,) * D))
 
         key = random.PRNGKey(0)
         key, subkey1, subkey2, subkey3, subkey4 = random.split(key, 5)
         L = 5
-        rand_layer = geom.BatchLayer(
+        rand_multi_image = geom.BatchMultiImage(
             {
                 (0, 0): random.normal(subkey1, shape=((L, 3) + (N,) * D)),
                 (1, 0): random.normal(subkey2, shape=((L, 1) + (N,) * D + (D,))),
@@ -762,9 +775,11 @@ class TestBatchLayer:
 
         layout = {(0, 0): 3, (1, 0): 1, (1, 1): 2, (2, 0): 1}
 
-        scalar_layer2 = rand_layer.to_scalar_layer()
-        assert list(scalar_layer2.keys()) == [(0, 0)]
-        assert rand_layer == rand_layer.to_scalar_layer().from_scalar_layer(layout)
+        scalar_multi_image2 = rand_multi_image.to_scalar_multi_image()
+        assert list(scalar_multi_image2.keys()) == [(0, 0)]
+        assert rand_multi_image == rand_multi_image.to_scalar_multi_image().from_scalar_multi_image(
+            layout
+        )
 
     def testTimesGroupElement(self):
         N = 5
@@ -777,12 +792,12 @@ class TestBatchLayer:
         )
         key = random.PRNGKey(0)
         for D in [2, 3]:
-            layer = geom.BatchLayer({}, D)
+            multi_image = geom.BatchMultiImage({}, D)
 
             for parity in [0, 1]:
                 for k in [0, 1, 2, 3]:
                     key, subkey = random.split(key)
-                    layer.append(
+                    multi_image.append(
                         k,
                         parity,
                         random.normal(subkey, shape=((batch, channels) + (N,) * D + (D,) * k)),
@@ -791,11 +806,11 @@ class TestBatchLayer:
             operators = geom.make_all_operators(D)
 
             for gg in operators:
-                rotated_layer = layer.times_group_element(gg)
+                rotated_multi_image = multi_image.times_group_element(gg)
 
-                for (k, parity), img_block in layer.items():
+                for (k, parity), img_block in multi_image.items():
                     rotated_block = vmap_times_gg(D, img_block, parity, gg)
-                    assert jnp.allclose(rotated_layer[(k, parity)], rotated_block)
+                    assert jnp.allclose(rotated_multi_image[(k, parity)], rotated_block)
 
     def testNorm(self):
         N = 5
@@ -807,7 +822,7 @@ class TestBatchLayer:
 
         # norm of scalars, pseudo scalars, and vectors
         key, subkey1, subkey2, subkey3 = random.split(key, 4)
-        layer = geom.BatchLayer(
+        multi_image = geom.BatchMultiImage(
             {
                 (0, 0): random.normal(subkey1, shape=(batch, channels) + (N,) * D),
                 (0, 1): random.normal(subkey2, shape=(batch, channels) + (N,) * D),
@@ -816,17 +831,17 @@ class TestBatchLayer:
             D,
         )
 
-        normed_layer = layer.norm()
-        assert list(normed_layer.keys()) == [(0, 0)]  # odd parity is converted to even parity
-        assert normed_layer[(0, 0)].shape == ((batch, 3 * channels) + (N,) * D)
-        assert jnp.allclose(normed_layer[(0, 0)][:, :channels], jnp.abs(layer[(0, 0)]))
+        normed_multi_image = multi_image.norm()
+        assert list(normed_multi_image.keys()) == [(0, 0)]  # odd parity is converted to even parity
+        assert normed_multi_image[(0, 0)].shape == ((batch, 3 * channels) + (N,) * D)
+        assert jnp.allclose(normed_multi_image[(0, 0)][:, :channels], jnp.abs(multi_image[(0, 0)]))
         assert jnp.allclose(
-            normed_layer[(0, 0)][:, channels : 2 * channels], jnp.abs(layer[(0, 1)])
+            normed_multi_image[(0, 0)][:, channels : 2 * channels], jnp.abs(multi_image[(0, 1)])
         )
         vector_norm = jnp.linalg.norm(
-            layer[(1, 0)].reshape(layer[(1, 0)].shape[: 2 + D] + (-1,)), axis=2 + D
+            multi_image[(1, 0)].reshape(multi_image[(1, 0)].shape[: 2 + D] + (-1,)), axis=2 + D
         )
-        assert jnp.allclose(normed_layer[(0, 0)][:, 2 * channels :], vector_norm)
+        assert jnp.allclose(normed_multi_image[(0, 0)][:, 2 * channels :], vector_norm)
 
     def testAdd(self):
         key = random.PRNGKey(time.time_ns())
@@ -839,7 +854,7 @@ class TestBatchLayer:
             key, 9
         )
 
-        layer1 = geom.BatchLayer(
+        multi_image1 = geom.BatchMultiImage(
             {
                 (0, 0): random.normal(subkey1, shape=((batch, channels) + (N,) * D + (D,) * 0)),
                 (1, 0): random.normal(subkey2, shape=((batch, channels) + (N,) * D + (D,) * 1)),
@@ -848,7 +863,7 @@ class TestBatchLayer:
             True,
         )
 
-        layer2 = geom.BatchLayer(
+        multi_image2 = geom.BatchMultiImage(
             {
                 (0, 0): random.normal(subkey3, shape=((batch, channels) + (N,) * D + (D,) * 0)),
                 (1, 0): random.normal(subkey4, shape=((batch, channels) + (N,) * D + (D,) * 1)),
@@ -857,60 +872,60 @@ class TestBatchLayer:
             True,
         )
 
-        layer3 = layer1 + layer2
-        layer4 = geom.BatchLayer(
+        multi_image3 = multi_image1 + multi_image2
+        multi_image4 = geom.BatchMultiImage(
             {
-                (0, 0): layer1[(0, 0)] + layer2[(0, 0)],
-                (1, 0): layer1[(1, 0)] + layer2[(1, 0)],
+                (0, 0): multi_image1[(0, 0)] + multi_image2[(0, 0)],
+                (1, 0): multi_image1[(1, 0)] + multi_image2[(1, 0)],
             },
             D,
             True,
         )
 
-        assert layer3 == layer4
+        assert multi_image3 == multi_image4
 
-        # mismatched layer types
-        layer5 = geom.BatchLayer(
+        # mismatched multi_image types
+        multi_image5 = geom.BatchMultiImage(
             {(0, 0): random.normal(subkey5, shape=((batch, channels) + (N,) * D + (D,) * 0))},
             D,
             True,
         )
-        layer6 = geom.BatchLayer(
+        multi_image6 = geom.BatchMultiImage(
             {(1, 0): random.normal(subkey6, shape=((batch, channels) + (N,) * D + (D,) * 1))},
             D,
             True,
         )
         with pytest.raises(AssertionError):
-            layer5 + layer6
+            multi_image5 + multi_image6
 
         # mismatched number of channels
-        layer7 = geom.Layer(
+        multi_image7 = geom.MultiImage(
             {(0, 0): random.normal(subkey7, shape=((batch, channels + 1) + (N,) * D + (D,) * 0))},
             D,
             True,
         )
-        layer8 = geom.Layer(
+        multi_image8 = geom.MultiImage(
             {(0, 0): random.normal(subkey8, shape=((batch, channels) + (N,) * D + (D,) * 0))},
             D,
             True,
         )
         with pytest.raises(TypeError):
-            layer7 + layer8
+            multi_image7 + multi_image8
 
         # mismatched batch size
         key, subkey9, subkey10 = random.split(key, 3)
-        layer9 = geom.Layer(
+        multi_image9 = geom.MultiImage(
             {(0, 0): random.normal(subkey9, shape=((batch + 1, channels) + (N,) * D + (D,) * 0))},
             D,
             True,
         )
-        layer10 = geom.Layer(
+        multi_image10 = geom.MultiImage(
             {(0, 0): random.normal(subkey10, shape=((batch, channels) + (N,) * D + (D,) * 0))},
             D,
             True,
         )
         with pytest.raises(TypeError):
-            layer9 + layer10
+            multi_image9 + multi_image10
 
     def testMul(self):
         key = random.PRNGKey(0)
@@ -921,7 +936,7 @@ class TestBatchLayer:
 
         key, subkey1, subkey2 = random.split(key, 3)
 
-        layer1 = geom.BatchLayer(
+        multi_image1 = geom.BatchMultiImage(
             {
                 (0, 0): random.normal(subkey1, shape=(batch, channels) + (N,) * D),
                 (1, 0): random.normal(subkey2, shape=(batch, channels) + (N,) * D + (D,)),
@@ -930,18 +945,18 @@ class TestBatchLayer:
             True,
         )
 
-        layer2 = layer1 * 3
-        assert jnp.allclose(layer2[(0, 0)], layer1[(0, 0)] * 3)
-        assert jnp.allclose(layer2[(1, 0)], layer1[(1, 0)] * 3)
-        assert layer2.D == D
-        assert layer2.is_torus == (True,) * D
+        multi_image2 = multi_image1 * 3
+        assert jnp.allclose(multi_image2[(0, 0)], multi_image1[(0, 0)] * 3)
+        assert jnp.allclose(multi_image2[(1, 0)], multi_image1[(1, 0)] * 3)
+        assert multi_image2.D == D
+        assert multi_image2.is_torus == (True,) * D
 
-        layer3 = layer1 * -1
-        assert jnp.allclose(layer3[(0, 0)], layer1[(0, 0)] * -1)
-        assert jnp.allclose(layer3[(1, 0)], layer1[(1, 0)] * -1)
-        assert layer2.D == D
-        assert layer2.is_torus == (True,) * D
+        multi_image3 = multi_image1 * -1
+        assert jnp.allclose(multi_image3[(0, 0)], multi_image1[(0, 0)] * -1)
+        assert jnp.allclose(multi_image3[(1, 0)], multi_image1[(1, 0)] * -1)
+        assert multi_image2.D == D
+        assert multi_image2.is_torus == (True,) * D
 
-        # try to multiply two layers together
+        # try to multiply two multi_images together
         with pytest.raises(AssertionError):
-            layer1 * layer1
+            multi_image1 * multi_image1
